renames project and adds more to initial page
