+++ index.js
@@ -32,7 +32,7 @@ const facetsWithArticles = require("./routes/facetsWithArticles");
 const heartbeat = require("./routes/heartbeat");
 const hierarchicalEdgeBundling = require("./routes/hierarchicalEdgeBundling");
 const tinder = require("./routes/tinder");
-const tabular = require("./routes/tabular");
+const ftabulous = require("./routes/ftabulous");
 
 const hbs = require("hbs");
 
@@ -78,7 +78,7 @@ app.use("/facetsWithArticles", facetsWithArticles);
 app.use("/heartbeat", heartbeat);
 app.use("/hierarchicalEdgeBundling", hierarchicalEdgeBundling);
 app.use("/tinder", tinder);
-app.use("/tabular", tabular);
+app.use("/ftabulous", ftabulous);
 
 // ---
