--- static/css/tabular.css
@@ -1,21 +1,0 @@
-body {
-	background: #fff1e5;
-	font-family: 'MetricWeb', sans-serif;
-}
-
-h1 {
-	font-family: 'FinancierDisplayWeb', serif;
-    text-transform: capitalize;
-    margin: 0 0 5px;
-    padding: 0;
-}
-
-h2 {
-	font-family: 'MetricWeb', sans-serif;
-    text-transform: capitalize;
-    font-size: 18px;
-    margin: 0 0 10px;
-    padding: 0;
-    font-weight: normal;
-    font-style: italic;
-}
