--- static/js/tabular/main.js
@@ -1,18 +1,0 @@
-//use strict 
-
-var data = null;
-
-
-function init(dataStr){
-	data = dataStr;
-
-	initControls();
-	addListeners();
-}
-
-function addListeners(){
-}
-
-function initControls(){
-}
-
