+++ views/ftabulous.hbs
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<title>FTabulous</title>
+		<link rel="shortcut icon" href="/static/img/icon/favicon.ico">
+		<link rel="stylesheet" href="https://build.origami.ft.com/v2/bundles/css?modules=o-fonts@^3.0.4,o-buttons@^5.10.3,o-forms@^5.2.4">
+		<link rel="stylesheet" type="text/css" href="/static/css/ftabulous.css">
+	</head>
+	<body>
+		<h1>FTabulous - Experimental dashboard</h1>
+		<h2>Using articles published in the last 24hrs</h2>
+		<div class="facetSelection">
+			<div class="inner">
+				<button type="submit" name="select" class="o-buttons o-buttons--primary" value="topics">Topics</button>
+				<button type="submit" name="select" class="o-buttons o-buttons--primary" value="people">People</button>
+				<button type="submit" name="select" class="o-buttons o-buttons--primary" value="organisations">Organisations</button>
+				<button type="submit" name="select" class="o-buttons o-buttons--primary" value="genre">Genre</button>
+			</div>
+		</div>
+		<script src="/static/js/ftabulous/main.js"></script>
+		<script>
+			init('{{data}}');
+		</script>
+	</body>
+</html>
