+++ views/index.hbs
@@ -44,7 +44,7 @@
 			<a href="tinder">FT as Tinder</a>
 		</li>
 		<li>
-			<a href="tabular">Tabular Data</a>
+			<a href="ftabulous">FTabulous</a>
 			<ul><li>Presenting summarised recent news in a table format</li></ul>
 		</li>
 	</ul>
