--- views/tabular.hbs
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<title>Tabular</title>
-		<link rel="shortcut icon" href="/static/img/icon/favicon.ico">
-		<link rel="stylesheet" href="https://build.origami.ft.com/v2/bundles/css?modules=o-fonts@^3.0.4,o-buttons@^5.10.3,o-forms@^5.2.4">
-		<link rel="stylesheet" type="text/css" href="/static/css/tabular.css">
-	</head>
-	<body>
-		<h1>Tabular - Demo Index Page</h1>
-		<h2>Using articles published in the last 24hrs</h2>
-		<script src="/static/js/tabular/main.js"></script>
-		<script>
-			init('{{data}}');
-		</script>
-	</body>
-</html>
